Lazy load inspector modules in Vega
